+++ router/router.go
@@ -95,8 +95,11 @@ func Configure(m *wserver.Manager, client remote.Client) *gin.Engine {
 		server.POST("/commands", postServerCommands)
 		server.POST("/install", postServerInstall)
 		server.POST("/reinstall", postServerReinstall)
+		server.POST("/import", postServerImport)
 		server.POST("/sync", postServerSync)
 		server.POST("/ws/deny", postServerDenyWSTokens)
+		server.POST("/proxy/create", postServerProxyCreate)
+		server.DELETE("/proxy/delete", postServerProxyDelete)
 
 		// This archive request causes the archive to start being created
 		// this should only be triggered by the panel.
@@ -109,6 +112,7 @@ func Configure(m *wserver.Manager, client remote.Client) *gin.Engine {
 		files := server.Group("/files")
 		{
 			files.GET("/contents", getServerFileContents)
+			files.GET("/fingerprints", getServerFileFingerprints)
 			files.GET("/list-directory", getServerListDirectory)
 			files.PUT("/rename", putServerRenameFiles)
 			files.POST("/copy", postServerCopyFile)